FIX: make plot in mono notebook "live"
When creating a Figure at least one message needs to make it from the
front end to the kernel.  However, the kernel can not process that
message until the cell you create in figure in has returned.  To make
the figure "live" in the notebook pre-create it with the agreed on name.