MNT: try to fix the networking
There is something wrong with multicast UDP search to multiple IOCs running on
the same (local) host so that it is basically round-robin which of the IOCs
will get the search so most of the time it will fail.

Unclear if this is a generic networking issue, a docker networking issue, or a
networking issue in this version of ubuntu.

This adjusts the flags to put each IOC on it's own IP address (apparently the
entire 127.0.0.* space is available for our use!) and then adds almost all of
those IPs to the EPICS_CA_ADDR_LIST env (so that if we add more IOCs we do not
have to remember to add things in two places).

Co-authored-by: Ken Lauer <klauer@users.noreply.github.com>